feat(#38): WorkerCheck with suppression
