+++ src/it/correct/src/main/java/User.java
@@ -1,5 +1,9 @@
-import ru.l3r8y.annotations.Worker;
-
-@Worker
-final class User {
-}
+/**
+ * @todo #38 Check @Worker annotation with invoker test.
+ *  Now annotation can not be compiled with invoker.
+ *  We need to compile it in order to to integration test
+ *  with WorkerCheck.java.
+ *  Don't forget to remove this puzzle.
+ */
+//final class User {
+//}
